% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAnanlysis.R
\name{edger_fun}
\alias{edger_fun}
\title{Differential Expression Analysis with edgeR}
\usage{
edger_fun(data, select = 1:4, group = NULL, LogFC = 0.5, spike = TRUE)
}
\arguments{
\item{data}{A matrix of raw counts.}

\item{select}{A vector of column indices to select from the data matrix (default: 1:4).}

\item{group}{A vector specifying group labels for the selected samples. Default: c("A", "A", "B", "B").}

\item{LogFC}{A numeric value for the log fold change threshold (default: 0.5).}

\item{spike}{A logical indicating if the spike in normalization has been done (default: TRUE).}
}
\value{
A data frame containing the results of the differential expression analysis, including logFC, FDR, and group labels ("UP", "DN", "NC").
}
\description{
This function performs differential expression analysis using the edgeR package.
}
\examples{
# Example usage
results <- edger_fun(data, select = 1:4, group = c("A", "A", "B", "B"))
}
