% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAnanlysis.R
\name{feature2count}
\alias{feature2count}
\title{Convert FeatureCount data into raw count matrix}
\usage{
feature2count(featureCount)
}
\arguments{
\item{featureCount}{A data frame containing FeatureCount data without the 1st row.}
}
\value{
A matrix with raw counts where row names are gene IDs.
}
\description{
Convert FeatureCount data into raw count matrix
}
\examples{
# Example usage
raw_count <- feature2count(featureCount)
}
