% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAnanlysis.R
\name{spike_norm}
\alias{spike_norm}
\title{Normalize data using size factors}
\usage{
spike_norm(data, sf)
}
\arguments{
\item{data}{A numeric matrix of raw counts.}

\item{sf}{A numeric vector of spike in reads.}
}
\value{
A normalized numeric matrix.
}
\description{
This function normalizes data using provided size factors.
}
\examples{
# Example usage
normalized_data <- spike_norm(data, sf)
}
